+++ backend/eventos/serializers/event.py
@@ -1,6 +1,5 @@
 from rest_framework import serializers
-from .models import Event
-
+from eventos.models.event import Event
 
 class EventSerializer(serializers.ModelSerializer):
     class Meta:
