+++ backend/user_auth/apps.py
@@ -1,6 +1,5 @@
 from django.apps import AppConfig
 
-
 class UserAuthConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'user_auth'
